frank-brain/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
openai-whisper==20231117
torch==2.1.0
torchaudio==2.1.0
python-dotenv==1.0.0
langchain-openai==0.0.2
langgraph==0.0.20